fix signed values shifts
Overflowing a signed value is undefined. 1 << 31 is also undefined on
signed int.

Try to use BIT macro, use UL suffix or cast to (u32) all shifted values
to prevent undefined behavior.

+++ u-boot/cpu/mips/ar7240/qca_ls_uart.c
@@ -60,7 +60,7 @@ int serial_init(void)
 	 * - parity: no
 	 */
 	qca_soc_reg_write(QCA_LSUART_LCR_REG,
-					  QCA_LSUART_LCR_CLS_8BIT_VAL << QCA_LSUART_LCR_CLS_SHIFT);
+					  (u32)QCA_LSUART_LCR_CLS_8BIT_VAL << QCA_LSUART_LCR_CLS_SHIFT);
 
 	return 0;
 }
